reorg io tests
